Add account status

+++ app.php
@@ -54,34 +54,38 @@ function doUpdate() {
 
 	// loop through accounts
 	foreach ($config['account'] as $t) {
-
-		// new twitter
-		$twitter = new Twitter($t['keys']['consumer_key'], $t['keys']['consumer_secret'], $t['keys']['access_token'], $t['keys']['access_token_secret']);
-
-		// grab the cam image
-		$ch = curl_init($t['webcam_url']);
-		$fp = fopen($config['tmp_img'], 'wb');
-		curl_setopt($ch, CURLOPT_FILE, $fp);
-		curl_setopt($ch, CURLOPT_HEADER, 0);
-		curl_exec($ch);
-		curl_close($ch);
-		fclose($fp);
-
-		// process hashtags
-		$tags = implode(' ', $t['hashtags']);
-
-		// set tweet bodycontent
-		$tweet_text = $t['tweet_text'] . ' ' . $tags;
-
-		try {
-			//tweet da tweet
-			$tweet = $twitter -> send($tweet_text, $config['tmp_img']);
-
-			// remove the temp image
-			unlink($config['tmp_img']);
-
-		} catch (TwitterException $e) {
-			echo 'Error: ' . $e -> getMessage();
+		
+		//if account is active
+		if($t['active']){
+			
+			// new twitter
+			$twitter = new Twitter($t['keys']['consumer_key'], $t['keys']['consumer_secret'], $t['keys']['access_token'], $t['keys']['access_token_secret']);
+	
+			// grab the cam image
+			$ch = curl_init($t['webcam_url']);
+			$fp = fopen($config['tmp_img'], 'wb');
+			curl_setopt($ch, CURLOPT_FILE, $fp);
+			curl_setopt($ch, CURLOPT_HEADER, 0);
+			curl_exec($ch);
+			curl_close($ch);
+			fclose($fp);
+	
+			// process hashtags
+			$tags = implode(' ', $t['hashtags']);
+	
+			// set tweet bodycontent
+			$tweet_text = $t['tweet_text'] . ' ' . $tags;
+	
+			try {
+				//tweet da tweet
+				$tweet = $twitter -> send($tweet_text, $config['tmp_img']);
+	
+				// remove the temp image
+				unlink($config['tmp_img']);
+	
+			} catch (TwitterException $e) {
+				echo 'Error: ' . $e -> getMessage();
+			}
 		}
 
 	}
